Move ingest command to guacingest
Remove neo4j options, make the "ingest" subcommand the default.
