+++ cmd/guacingest/cmd/root.go
@@ -31,11 +31,6 @@ import (
 var cfgFile string
 
 var flags = struct {
-	dbAddr  string
-	gdbuser string
-	gdbpass string
-	realm   string
-
 	// nats
 	natsAddr string
 
@@ -49,14 +44,10 @@ var flags = struct {
 func init() {
 	cobra.OnInitialize(initConfig)
 	persistentFlags := rootCmd.PersistentFlags()
-	persistentFlags.StringVar(&flags.dbAddr, "gdbaddr", "neo4j://localhost:7687", "address to neo4j db")
-	persistentFlags.StringVar(&flags.gdbuser, "gdbuser", "", "neo4j user credential to connect to graph db")
-	persistentFlags.StringVar(&flags.gdbpass, "gdbpass", "", "neo4j password credential to connect to graph db")
-	persistentFlags.StringVar(&flags.realm, "realm", "neo4j", "realm to connect to graph db")
 	persistentFlags.StringVar(&flags.natsAddr, "natsaddr", "nats://127.0.0.1:4222", "address to connect to NATs Server")
 	persistentFlags.StringVar(&flags.collectSubAddr, "csub-addr", "localhost:2782", "address to connect to collect-sub service")
 	persistentFlags.StringVar(&flags.graphqlEndpoint, "gql-endpoint", "http://localhost:8080/query", "endpoint used to connect to graphQL server")
-	flagNames := []string{"gdbaddr", "gdbuser", "gdbpass", "realm", "natsaddr", "csub-addr", "gql-endpoint"}
+	flagNames := []string{"natsaddr", "csub-addr", "gql-endpoint"}
 	for _, name := range flagNames {
 		if flag := persistentFlags.Lookup(name); flag != nil {
 			if err := viper.BindPFlag(name, flag); err != nil {
@@ -99,8 +90,11 @@ func initConfig() {
 }
 
 var rootCmd = &cobra.Command{
-	Use:   "ingestor",
-	Short: "ingestor is an ingestor cmdline for GUAC",
+	Use:   "guacingest",
+	Short: "starts the GUAC processor, ingestor and assembler process",
+	Run: func(cmd *cobra.Command, args []string) {
+		ingest(cmd, args)
+	},
 }
 
 func Execute() {
